Revert "IntegerMod: new class field for value of phi(modulo)"
This reverts commit 42ef221fab03f8c218f1a1c5c0e3e4c4a9e815c1.

We wrote in that commit log:
 ``When declaring an IntegerMod subclass, it is now possible to
   optionally specify the value of the phi Euler function of
   the modulo; this makes it possible to efficiently compute
   exponentiation (mod m) for even very large values of the
   exponent, since it's well known that a^phi(m) = 1 (mod m)''

But that's false, as a^phi(m) = 1 (mod m) iff gcd(a, m) = 1.

Moreover, we really won't really need efficient exponentiation
for big exponents, since in RSA we need only to be able to
calculate x^e (mod n) for x < n and e < phi(n).